feat: parking request

+++ parkingmanagement/src/main/java/com/parkingmanagement/parkingmanagement/controller/ParkingController.java
@@ -1,20 +1,27 @@
 package com.parkingmanagement.parkingmanagement.controller;
 
+import com.parkingmanagement.parkingmanagement.model.dto.ParkingDTO;
 import com.parkingmanagement.parkingmanagement.model.entity.Parking;
 import com.parkingmanagement.parkingmanagement.service.ParkingService;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.parkingmanagement.parkingmanagement.service.UserService;
+import jakarta.validation.Valid;
+import org.springframework.beans.BeanUtils;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-import java.util.UUID;
+import java.util.*;
 
 @RestController
 @RequestMapping("/parking")
 public class ParkingController {
 
-    @Autowired
-    private ParkingService parkingService;
+    private final ParkingService parkingService;
+    private final UserService userService;
+
+    public ParkingController(ParkingService parkingService, UserService userService) {
+        this.parkingService = parkingService;
+        this.userService = userService;
+    }
 
     @GetMapping("/{id}")
     public ResponseEntity<Parking> getById(@PathVariable UUID id) {
@@ -23,13 +30,42 @@ public class ParkingController {
                 .orElseGet(() -> ResponseEntity.notFound().build());
     }
 
-    @PostMapping
-    public Parking create(@RequestBody Parking parking) {
-        return parkingService.save(parking);
+    @GetMapping("/user/{userId}")
+    public ResponseEntity<List<Parking>> getByUserId(@PathVariable UUID userId) {
+        List<Parking> parkings = new ArrayList<>();
+        parkings.addAll(parkingService.findByUserCreatorId(userId));
+        parkings.addAll(parkingService.findByEmployeeUserId(userId));
+
+        // Remove objetos iguais
+        Set<Parking> uniqueParkings = new HashSet<>(parkings);
+        parkings.clear();
+        parkings.addAll(uniqueParkings);
+
+        return ResponseEntity.ok(parkings);
+    }
+
+    @PostMapping("/register")
+    public ResponseEntity<Parking> register(@Valid @RequestBody ParkingDTO parkingDTO) {
+        if (!userService.existsById(UUID.fromString(parkingDTO.getUserCreatorId()))) {
+            return ResponseEntity.badRequest().build();
+        }
+
+        Parking newParking = new Parking();
+        BeanUtils.copyProperties(parkingDTO, newParking);
+
+        Parking savedParking = parkingService.save(newParking);
+
+        return ResponseEntity.ok(savedParking);
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable UUID id) {
+    public ResponseEntity<Object> delete(@PathVariable UUID id) {
+        if (!parkingService.existsById(id)) {
+            return ResponseEntity.notFound().build();
+        }
+
         parkingService.delete(id);
+
+        return ResponseEntity.ok().build();
     }
 }
